Fix refresh loseing the member
